+++ src-console/ConsoleAppEF2.1.1/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Linq.Dynamic.Core;
 using System.Linq.Dynamic.Core.CustomTypeProviders;
@@ -7,7 +8,6 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using ConsoleAppEF2.Database;
-using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 
@@ -19,13 +19,19 @@ namespace ConsoleAppEF211
         {
             public HashSet<Type> GetCustomTypes()
             {
-                var assemblies = AppDomain.CurrentDomain.GetAssemblies();
+                // https://stackoverflow.com/a/2384679/255966
+                var loadedAssemblies = AppDomain.CurrentDomain.GetAssemblies().ToList();
+                var loadedPaths = loadedAssemblies.Where(a => !a.IsDynamic).Select(a => a.Location).ToArray();
 
-                var set = new HashSet<Type>(FindTypesMarkedWithDynamicLinqTypeAttribute(assemblies));
+                var referencedPaths = Directory.GetFiles(AppDomain.CurrentDomain.BaseDirectory, "*.dll");
+                var toLoad = referencedPaths.Where(r => !loadedPaths.Contains(r, StringComparer.InvariantCultureIgnoreCase)).ToList();
 
-                set.Add(typeof(TestContext));
+                toLoad.ForEach(path => loadedAssemblies.Add(AppDomain.CurrentDomain.Load(AssemblyName.GetAssemblyName(path))));
 
-                return set;
+                return new HashSet<Type>(FindTypesMarkedWithDynamicLinqTypeAttribute(loadedAssemblies))
+                {
+                    typeof(TestContext)
+                };
             }
 
             public Dictionary<Type, List<MethodInfo>> GetExtensionMethods()
@@ -103,6 +109,7 @@ namespace ConsoleAppEF211
             Console.WriteLine("all {0}", JsonConvert.SerializeObject(all, Formatting.Indented));
 
             var config = ParsingConfig.DefaultEFCore21;
+            //config.ResolveTypesBySimpleName = true;
             config.CustomTypeProvider = new C();
 
             var context = new TestContext();