+++ src/System.Linq.Dynamic.Core/Parser/ExpressionParser.cs
@@ -904,10 +904,9 @@ namespace System.Linq.Dynamic.Core.Parser
                 }
                 else
                 {
-                    var lambda = expr as LambdaExpression;
-                    if (lambda != null)
+                    if (expr is LambdaExpression lambdaExpression)
                     {
-                        return ParseLambdaInvocation(lambda);
+                        return ParseLambdaInvocation(lambdaExpression);
                     }
                 }
 
@@ -1021,14 +1020,25 @@ namespace System.Linq.Dynamic.Core.Parser
 
             Expression[] args = ParseArgumentList();
 
-            if (args.Length != 1)
+            if (args.Length != 1 && args.Length != 2)
             {
-                throw ParseError(errorPos, Res.FunctionRequiresOneArg, functionName);
+                throw ParseError(errorPos, Res.FunctionRequiresOneOrTwoArgs, functionName);
             }
 
-            Type resolvedType = ResolveTypeFromArgumentExpression(functionName, args[0]);
+            Expression typeArgument;
+            Expression it;
+            if (args.Length == 1)
+            {
+                typeArgument = args[0];
+                it = _it;
+            }
+            else
+            {
+                typeArgument = args[1];
+                it = args[0];
+            }
 
-            return Expression.TypeIs(_it, resolvedType);
+            return Expression.TypeIs(it, ResolveTypeFromArgumentExpression(functionName, typeArgument, args.Length));
         }
 
         // As(...) function
@@ -1040,14 +1050,25 @@ namespace System.Linq.Dynamic.Core.Parser
 
             Expression[] args = ParseArgumentList();
 
-            if (args.Length != 1)
+            if (args.Length != 1 && args.Length != 2)
             {
-                throw ParseError(errorPos, Res.FunctionRequiresOneArg, functionName);
+                throw ParseError(errorPos, Res.FunctionRequiresOneOrTwoArgs, functionName);
             }
 
-            Type resolvedType = ResolveTypeFromArgumentExpression(functionName, args[0]);
+            Expression typeArgument;
+            Expression it;
+            if (args.Length == 1)
+            {
+                typeArgument = args[0];
+                it = _it;
+            }
+            else
+            {
+                typeArgument = args[1];
+                it = args[0];
+            }
 
-            return Expression.TypeAs(_it, resolvedType);
+            return Expression.TypeAs(it, ResolveTypeFromArgumentExpression(functionName, typeArgument, args.Length));
         }
 
         // Cast(...) function
@@ -1059,14 +1080,25 @@ namespace System.Linq.Dynamic.Core.Parser
 
             Expression[] args = ParseArgumentList();
 
-            if (args.Length != 1)
+            if (args.Length != 1 && args.Length != 2)
             {
-                throw ParseError(errorPos, Res.FunctionRequiresOneArg, functionName);
+                throw ParseError(errorPos, Res.FunctionRequiresOneOrTwoArgs, functionName);
             }
 
-            Type resolvedType = ResolveTypeFromArgumentExpression(functionName, args[0]);
+            Expression typeArgument;
+            Expression it;
+            if (args.Length == 1)
+            {
+                typeArgument = args[0];
+                it = _it;
+            }
+            else
+            {
+                typeArgument = args[1];
+                it = args[0];
+            }
 
-            return Expression.ConvertChecked(_it, resolvedType);
+            return Expression.ConvertChecked(it, ResolveTypeFromArgumentExpression(functionName, typeArgument, args.Length));
         }
 
         Expression GenerateConditional(Expression test, Expression expressionIfTrue, Expression expressionIfFalse, bool nullPropagating, int errorPos)
@@ -1101,8 +1133,7 @@ namespace System.Linq.Dynamic.Core.Parser
                     {
                         // - create nullable constant from expressionIfTrue with type from expressionIfFalse
                         // - convert expressionIfFalse to nullable (unless it's already nullable)
-
-                        Type nullableType = TypeHelper.ToNullableType(expressionIfFalse.Type);
+                        var nullableType = TypeHelper.ToNullableType(expressionIfFalse.Type);
                         expressionIfTrue = Expression.Constant(null, nullableType);
 
                         if (!TypeHelper.IsNullableType(expressionIfFalse.Type))
@@ -1343,24 +1374,25 @@ namespace System.Linq.Dynamic.Core.Parser
 #endif
             }
 
-            IEnumerable<PropertyInfo> propertyInfos = type.GetProperties();
+            // Option 1. Try to bind via properties (TODO : investigate if this code block is 100% correct and is needed) 
+            var propertyInfos = type.GetProperties();
             if (type.GetTypeInfo().BaseType == typeof(DynamicClass))
             {
-                propertyInfos = propertyInfos.Where(x => x.Name != "Item");
+                propertyInfos = propertyInfos.Where(x => x.Name != "Item").ToArray();
             }
-
-            Type[] propertyTypes = propertyInfos.Select(p => p.PropertyType).ToArray();
-            ConstructorInfo ctor = type.GetConstructor(propertyTypes);
+            var propertyTypes = propertyInfos.Select(p => p.PropertyType).ToArray();
+            var ctor = type.GetConstructor(propertyTypes);
             if (ctor != null)
             {
-                var ctorParameters = ctor.GetParameters();
-                if (ctorParameters.Length == expressions.Count)
+                var constructorParameters = ctor.GetParameters();
+                if (constructorParameters.Length == expressions.Count)
                 {
-                    bool bindParametersSequentially = !properties.All(p => ctorParameters
+                    bool bindParametersSequentially = !properties.All(p => constructorParameters
                         .Any(cp => cp.Name == p.Name && (cp.ParameterType == p.Type || p.Type == Nullable.GetUnderlyingType(cp.ParameterType))));
                     var expressionsPromoted = new List<Expression>();
+
                     // Loop all expressions and promote if needed
-                    for (int i = 0; i < ctorParameters.Length; i++)
+                    for (int i = 0; i < constructorParameters.Length; i++)
                     {
                         if (bindParametersSequentially)
                         {
@@ -1368,8 +1400,8 @@ namespace System.Linq.Dynamic.Core.Parser
                         }
                         else
                         {
-                            Type propertyType = ctorParameters[i].ParameterType;
-                            string cParameterName = ctorParameters[i].Name;
+                            Type propertyType = constructorParameters[i].ParameterType;
+                            string cParameterName = constructorParameters[i].Name;
                             var propertyAndIndex = properties.Select((p, index) => new { p, index })
                                 .First(p => p.p.Name == cParameterName && (p.p.Type == propertyType || p.p.Type == Nullable.GetUnderlyingType(propertyType)));
                             // Promote from Type to Nullable Type if needed
@@ -1381,8 +1413,22 @@ namespace System.Linq.Dynamic.Core.Parser
                 }
             }
 
-            MemberBinding[] bindings = new MemberBinding[properties.Count];
-            for (int i = 0; i < bindings.Length; i++)
+            // Option 2. Try to find a constructor with the exact argument-types and exact same order
+            var constructorArgumentTypes = properties.Select(p => p.Type).ToArray();
+            var exactConstructor = type.GetConstructor(constructorArgumentTypes);
+            if (exactConstructor != null)
+            {
+                // Promote from Type to Nullable Type if needed
+                var expressionsPromoted = exactConstructor.GetParameters()
+                    .Select((t, i) => _parsingConfig.ExpressionPromoter.Promote(expressions[i], t.ParameterType, true, true))
+                    .ToArray();
+
+                return Expression.New(exactConstructor, expressionsPromoted);
+            }
+
+            // Option 2. Call the default (empty) constructor and set the members
+            var memberBindings = new MemberBinding[properties.Count];
+            for (int i = 0; i < memberBindings.Length; i++)
             {
                 string propertyOrFieldName = properties[i].Name;
                 Type propertyOrFieldType;
@@ -1406,10 +1452,10 @@ namespace System.Linq.Dynamic.Core.Parser
                 }
 
                 // Promote from Type to Nullable Type if needed
-                bindings[i] = Expression.Bind(memberInfo, _parsingConfig.ExpressionPromoter.Promote(expressions[i], propertyOrFieldType, true, true));
+                memberBindings[i] = Expression.Bind(memberInfo, _parsingConfig.ExpressionPromoter.Promote(expressions[i], propertyOrFieldType, true, true));
             }
 
-            return Expression.MemberInit(Expression.New(type), bindings);
+            return Expression.MemberInit(Expression.New(type), memberBindings);
         }
 
         Expression ParseLambdaInvocation(LambdaExpression lambda)
@@ -1493,8 +1539,6 @@ namespace System.Linq.Dynamic.Core.Parser
                         throw ParseError(errorPos, Res.AmbiguousConstructorInvocation, TypeHelper.GetTypeName(type));
                 }
             }
-
-            // throw ParseError(errorPos, Res.CannotConvertValue, TypeHelper.GetTypeName(exprType), TypeHelper.GetTypeName(type));
 
             _textParser.ValidateToken(TokenId.Dot, Res.DotOrOpenParenOrStringLiteralExpected);
             _textParser.NextToken();
@@ -1709,7 +1753,7 @@ namespace System.Linq.Dynamic.Core.Parser
                 }
             }
 
-            var enumTypeAsString = string.Join("", parts.Take(parts.Count - 2).ToArray());
+            var enumTypeAsString = string.Concat(parts.Take(parts.Count - 2).ToArray());
             var enumType = _typeFinder.FindTypeByName(enumTypeAsString, null, true);
             if (enumType == null)
             {
@@ -1836,8 +1880,9 @@ namespace System.Linq.Dynamic.Core.Parser
             return Expression.Call(callType, methodName, typeArgs, args);
         }
 
-        private Type ResolveTypeFromArgumentExpression(string functionName, Expression argumentExpression)
+        private Type ResolveTypeFromArgumentExpression(string functionName, Expression argumentExpression, int? arguments = null)
         {
+            string argument = arguments == null ? string.Empty : arguments == 1 ? "first " : "second ";
             switch (argumentExpression)
             {
                 case ConstantExpression constantExpression:
@@ -1850,21 +1895,16 @@ namespace System.Linq.Dynamic.Core.Parser
                             return type;
 
                         default:
-                            throw ParseError(_textParser.CurrentToken.Pos, Res.FunctionRequiresOneNotNullArgOfType, functionName, "string or System.Type");
+                            throw ParseError(_textParser.CurrentToken.Pos, Res.FunctionRequiresNotNullArgOfType, functionName, argument, "string or System.Type");
                     }
 
                 default:
-                    throw ParseError(_textParser.CurrentToken.Pos, Res.FunctionRequiresOneNotNullArgOfType, functionName, "ConstantExpression");
+                    throw ParseError(_textParser.CurrentToken.Pos, Res.FunctionRequiresNotNullArgOfType, functionName, argument, "ConstantExpression");
             }
         }
 
         private Type ResolveTypeStringFromArgument(string functionName, string typeName)
         {
-            if (string.IsNullOrEmpty(typeName))
-            {
-                throw ParseError(_textParser.CurrentToken.Pos, Res.FunctionRequiresOneNotNullArg, functionName, typeName);
-            }
-
             Type resultType = _typeFinder.FindTypeByName(typeName, new[] { _it, _parent, _root }, true);
             if (resultType == null)
             {
@@ -1938,7 +1978,16 @@ namespace System.Linq.Dynamic.Core.Parser
                     throw ParseError(errorPos, Res.NoApplicableIndexer,
                         TypeHelper.GetTypeName(expr.Type));
                 case 1:
-                    return Expression.Call(expr, (MethodInfo)mb, args);
+                    var indexMethod = (MethodInfo)mb;
+                    var indexParameterType = indexMethod.GetParameters().First().ParameterType;
+
+                    var indexArgumentExpression = args[0]; // Indexer only has 1 parameter, so we can use args[0] here
+                    if (indexParameterType != indexArgumentExpression.Type)
+                    {
+                        indexArgumentExpression = Expression.Convert(indexArgumentExpression, indexParameterType);
+                    }
+
+                    return Expression.Call(expr, indexMethod, indexArgumentExpression);
 
                 default:
                     throw ParseError(errorPos, Res.AmbiguousIndexerInvocation, TypeHelper.GetTypeName(expr.Type));
@@ -1962,7 +2011,7 @@ namespace System.Linq.Dynamic.Core.Parser
             var memberExpression = expression as MemberExpression;
             if (memberExpression == null && expression.NodeType == ExpressionType.Coalesce)
             {
-                memberExpression = (expression as BinaryExpression).Left as MemberExpression;
+                memberExpression = (expression as BinaryExpression)?.Left as MemberExpression;
             }
 
             if (memberExpression != null)