+++ test/System.Linq.Dynamic.Core.Tests/Entities/EmployeeWrapper.cs
@@ -1,0 +1,7 @@
+﻿namespace System.Linq.Dynamic.Core.Tests.Entities
+{
+    public class EmployeeWrapper
+    {
+        public BaseEmployee Employee { get; set; }
+    }
+}