+++ test/System.Linq.Dynamic.Core.Tests/QueryableTests.cs
@@ -1,7 +1,4 @@
-﻿using System.Linq.Dynamic.Core.Tests.Helpers.Models;
-using Xunit;
-
-namespace System.Linq.Dynamic.Core.Tests
+﻿namespace System.Linq.Dynamic.Core.Tests
 {
     public partial class QueryableTests
     {